requirements.txt
torch
transformers
datasets
sentencepiece
accelerate
faiss-cpu
lightgbm
scikit-learn
numpy
pandas
tqdm
fastapi
uvicorn
mlflow
rank_bm25
python-dotenv
pyyaml
pyarrow
ruff
pytest
